[Refactor] ImageHandler 클래스명 변경 #119
