+++ src/main/java/com/yzgeneration/evc/domain/image/implement/ImageHandler.java
@@ -1,0 +1,6 @@
+package com.yzgeneration.evc.domain.image.implement;
+
+public interface ImageHandler {
+    String getPresignedURLForUpload(String prefix, String fileName);
+    String getImageUrl(String fileName);
+}
